move interactive experience to interactive dir in dotnet-scaffold
