+++ src/dotnet-scaffolding/dotnet-scaffold/Interactive/AppBuilder/ScaffoldCommandApp.cs
@@ -1,6 +1,7 @@
-using Microsoft.DotNet.Tools.Scaffold.Command;
+using Microsoft.DotNet.Tools.Scaffold.Interactive.Command;
+using Spectre.Console.Cli;
 
-namespace Microsoft.DotNet.Tools.Scaffold;
+namespace Microsoft.DotNet.Tools.Scaffold.Interactive.AppBuilder;
 
 /// <summary>
 /// Represents the application entry point for running the scaffold command.
@@ -12,14 +13,14 @@ internal class ScaffoldCommandApp
     private readonly string[] _args;
 
     // The Spectre.Console command application instance for the scaffold command.
-    private readonly Spectre.Console.Cli.CommandApp<ScaffoldCommand> _commandApp;
+    private readonly CommandApp<ScaffoldCommand> _commandApp;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="ScaffoldCommandApp"/> class.
     /// </summary>
     /// <param name="commandApp">The command application to execute.</param>
     /// <param name="args">The command-line arguments.</param>
-    public ScaffoldCommandApp(Spectre.Console.Cli.CommandApp<ScaffoldCommand> commandApp, string[] args)
+    public ScaffoldCommandApp(CommandApp<ScaffoldCommand> commandApp, string[] args)
     {
         _commandApp = commandApp;
         _args = args;
