+++ src/dotnet-scaffolding/dotnet-scaffold/Interactive/Command/ToolListSettings.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.DotNet.Tools.Scaffold.Command;
+namespace Microsoft.DotNet.Tools.Scaffold.Interactive.Command;
 
 /// <summary>
 /// Represents the settings used for listing installed tools via the scaffold CLI.
