+++ src/dotnet-scaffolding/dotnet-scaffold/Interactive/Command/ToolUninstallSettings.cs
@@ -3,7 +3,7 @@
 
 using Spectre.Console.Cli;
 
-namespace Microsoft.DotNet.Tools.Scaffold.Command;
+namespace Microsoft.DotNet.Tools.Scaffold.Interactive.Command;
 
 /// <summary>
 /// Represents the settings used for uninstalling a tool via the scaffold CLI.
