+++ src/dotnet-scaffolding/dotnet-scaffold/Interactive/Flow/FlowContextProperties.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.DotNet.Tools.Scaffold.Flow;
+namespace Microsoft.DotNet.Tools.Scaffold.Interactive.Flow;
 
 /// <summary>
 /// Provides constant property names used for storing and retrieving values in the flow context during scaffolding operations.
