+++ src/dotnet-scaffolding/dotnet-scaffold/Interactive/Flow/Steps/ComponentDiscovery.cs
@@ -4,7 +4,7 @@ using Microsoft.DotNet.Scaffolding.Core.ComponentModel;
 using Microsoft.DotNet.Tools.Scaffold.Services;
 using Spectre.Console.Flow;
 
-namespace Microsoft.DotNet.Tools.Scaffold.Flow.Steps;
+namespace Microsoft.DotNet.Tools.Scaffold.Interactive.Flow.Steps;
 
 /// <summary>
 /// Handles the discovery and selection of a scaffolding component (dotnet tool).
