+++ src/dotnet-scaffolding/dotnet-scaffold/Interactive/Flow/Steps/ParameterDiscovery.cs
@@ -10,7 +10,7 @@ using Microsoft.DotNet.Scaffolding.Roslyn.Services;
 using Spectre.Console;
 using Spectre.Console.Flow;
 
-namespace Microsoft.DotNet.Tools.Scaffold.Flow.Steps
+namespace Microsoft.DotNet.Tools.Scaffold.Interactive.Flow.Steps
 {
     /// <summary>
     /// Handles the discovery and user input for a parameter, including interactive pickers and validation.
