+++ src/dotnet-scaffolding/dotnet-scaffold/Interactive/Services/IFlowProvider.cs
@@ -3,7 +3,7 @@
 
 using Spectre.Console.Flow;
 
-namespace Microsoft.DotNet.Tools.Scaffold.Services;
+namespace Microsoft.DotNet.Tools.Scaffold.Interactive.Services;
 
 /// <summary>
 /// A factory interface for creating and managing interactive flows in the CLI.
